Fixed regression in behaviors resulting from failures in sync_notify.
Apparently using "redirect" does not work from within callbacks (per the device logs). This was causing all of the failure scenarios in sync_notify to fail to redirect to the login screen.

Need to be careful when using redirect/render; should never be used in callbacks as the controller is not in a rendering/redirect context in those scenarios.