Changed parent logic in activity model to use our custom find logic; the old methodology required the presence of the opportunityid field, which is only populated if the opportunity in question was created in CRM.
If an opportunity is created on the device on a slow connection/no connection, it was possible that an appointment could get created for that opportunity but the appointments tab couldn't resolve the appointment's parent opportunity as it relied on the opportunityid field being populated on the opportunity (which only happens coincidentally).

We should never rely on secondary ID fields (opportunityid, contactid, activityid, etc.), as these are not guaranteed to be populated. We should ALWAYS use the custom find logic which leverages the object and temp_id columns to uniquely identify an object (no matter where it was created).

Fixes defect 3952.