Fix for infinite sync loop between AppInfo & SearchContacts.
If a user is able to login to the app but gets a 401 unauthorized when querying CRM (sometimes old credentials work with login but not with querying), then the app would get in a loop where it would try to login in the background and then trigger another sync, which would fail and then login and trigger another sync, and so forth.

A full solution will likely require a fix from Rhomobile in RhoSync that will keep trying to sync on every request until successful.